cloud_run/ingest_players_job/requirements.txt
requests==2.32.3
google-cloud-logging==3.11.3